crowdin: sync: update push step
I've updated my remotes accordingly, this __should__ work perfectly!
All non gerrit repositories using this need to be added to the list, or the translations will get
pushed to refs/for/pie branch of the GitHub repository, when it needs to go to pie (refs/heads/pie)

Also update path to XML, we have a separate one for crowdin now!

Change-Id: Ief2330c525f8a1b944bd708d70d362dc8370ea6b
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>